+++ next.config.js
@@ -1,7 +1,7 @@
-const withTM = require("next-transpile-modules")(["react-konva", "konva"]);
-
-module.exports = withTM({
-  images: {
-    domains: ["firebasestorage.googleapis.com", "storage.googleapis.com"],
-  },
-});
+module.exports = (phase, { defaultConfig }) => {
+  return {
+    images: {
+      domains: ["firebasestorage.googleapis.com", "storage.googleapis.com"],
+    },
+  };
+};
